[AGDXML Parsing] Grid (Definitions)

+++ absGDX-framework/src/de/samdev/absgdx/framework/menu/agdxml/AgdxmlGridDefinitionsValue.java
@@ -2,10 +2,10 @@ package de.samdev.absgdx.framework.menu.agdxml;
 
 public class AgdxmlGridDefinitionsValue {
 
-	public final int value;
+	public final float value;
 	public final AgdxmlGridDefinitionsUnit unit;
 	
-	public AgdxmlGridDefinitionsValue(int value, AgdxmlGridDefinitionsUnit unit) {
+	public AgdxmlGridDefinitionsValue(float value, AgdxmlGridDefinitionsUnit unit) {
 		this.value = value;
 		this.unit = unit;
 	}
